requirements.txt
PyYAML>=6.0
schedule>=1.1.0
opencv-python>=4.6.0
google-api-python-client>=2.33.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.2
boto3>=1.20.0
requests>=2.25.0